Add 10sec timeout to openssl
When trying to parse URLs from a list which consists of old/non-existant or not reachable urls, openssl will hang forever. Therefore and for slower connections a 10sec timeout is very useful in our usecase, so I propose this addition here for others. Maybe not the most elegant solution but it works great.